+++ etl-service/src/main/java/com/sitefilm/etl/dto/DataPersonTranslation.java
@@ -3,7 +3,7 @@ package com.sitefilm.etl.dto;
 import java.time.OffsetDateTime;
 
 public record DataPersonTranslation(
-        Integer externalPersonId,
+        Long externalPersonId,
         String locale,
         String localeName,
         String biography
