+++ etl-service/src/main/java/com/sitefilm/etl/entity/enums/ContentStatus.java
@@ -7,20 +7,20 @@ import java.util.stream.Stream;
 
 @Getter
 public enum ContentStatus {
-    RUMORED(1, "Rumored"),
-    PLANNED(2, "Planned"),
-    IN_PRODUCTION(3, "In Production"),
-    POST_PRODUCTION(4, "Post Production"),
-    RELEASED(5, "Released"),
-    CANCELED(6, "Canceled"),
-    RETURNING_SERIES(7, "Returning Series"),
-    ENDED(8, "Ended"),
-    PILOT(9, "Pilot");
+    RUMORED((short) 1, "Rumored"),
+    PLANNED((short) 2, "Planned"),
+    IN_PRODUCTION((short) 3, "In Production"),
+    POST_PRODUCTION((short) 4, "Post Production"),
+    RELEASED((short) 5, "Released"),
+    CANCELED((short) 6, "Canceled"),
+    RETURNING_SERIES((short) 7, "Returning Series"),
+    ENDED((short) 8, "Ended"),
+    PILOT((short) 9, "Pilot");
 
-    private final int id;
+    private final Short id;
     private final String status;
 
-    ContentStatus(int id, String status) {
+    ContentStatus(Short id, String status) {
         this.id = id;
         this.status = status;
     }
