+++ etl-service/src/main/java/com/sitefilm/etl/service/core/dto/ContentCountryPersistDto.java
@@ -1,7 +1,7 @@
 package com.sitefilm.etl.service.core.dto;
 
 public record ContentCountryPersistDto(
-        Integer externalContentId,
-        Long countryId
+        Long externalContentId,
+        Short countryId
 ) {
 }
